+++ src/components/TabNavigation/TabNavigation.tsx
@@ -1,9 +1,18 @@
 import React from 'react';
 import styled, { css } from 'styled-components';
 import { ifProp } from 'styled-tools';
-import { AnalyticsProps } from '../../analytics';
+import {
+  AnalyticsComponent,
+  AnalyticsComponentType,
+  AnalyticsEvents,
+  AnalyticsProps,
+} from '../../analytics';
 
-const Tab = styled.a<{ active: boolean }>`
+type TabProps = {
+  active: boolean;
+} & AnalyticsProps;
+
+const Tab = styled.a<TabProps>`
   height: 100%;
   display: flex;
   align-items: center;
@@ -40,6 +49,7 @@ const TabContainer = styled.div`
 
 type NavigationTab = {
   title: string;
+  subTag?: string;
 } & AnalyticsProps;
 
 export type Props = {
@@ -51,14 +61,28 @@ export type Props = {
 const TabNavigation = ({ tabs, onTabClick, selectedTabIndex }: Props) => (
   <TabContainer>
     {tabs.map((tab, index) => (
-      <Tab
-        key={tab.title}
-        id={tab.title.toLowerCase()}
-        onClick={() => onTabClick(index)}
-        active={selectedTabIndex === index}
-      >
-        {tab.title}
-      </Tab>
+      <AnalyticsComponent key={tab.title}>
+        {({ pages, tracker, merchantId, ip }) => (
+          <Tab
+            key={tab.title}
+            id={tab.title.toLowerCase()}
+            onClick={() => {
+              tracker.track({
+                event: AnalyticsEvents.CLICK,
+                type: AnalyticsComponentType.TAB,
+                tag: tab.subTag || tab.tag || '',
+                pages,
+                merchantId,
+                ip,
+              });
+              onTabClick(index);
+            }}
+            active={selectedTabIndex === index}
+          >
+            {tab.title}
+          </Tab>
+        )}
+      </AnalyticsComponent>
     ))}
   </TabContainer>
 );