+++ src/components/Tabs/Tab.tsx
@@ -1,10 +1,12 @@
 import styled, { css } from 'styled-components';
 import { ifProp } from 'styled-tools';
-
-import { AnalyticsProps } from '../../analytics';
 import colors from '../../theme/colors';
 
-const TabWrapper = styled.div<{ selected: boolean }>`
+interface Props {
+  selected: boolean;
+}
+
+const Tab = styled.div<Props>`
   margin: 0;
   padding: 0.5rem 0;
   display: flex;
@@ -42,11 +44,5 @@ const TabWrapper = styled.div<{ selected: boolean }>`
     `
   )};
 `;
-
-export type Props = {
-  selected: boolean;
-} & AnalyticsProps;
-
-const Tab = ({ tag, ...rest }: Props) => <TabWrapper {...rest} />
 
 export default Tab;